reuse Commands type and collect_commands! macro
